+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - remove a default spacy abbreviation (`ts.`)
+- option for max scope on qualifier rules, limiting the number of tokens it applies to
 
 ### Fixed
 
 - issue where entity and context trigger were overlapping (e.g. `geen eetlust`)
+- some tests that were not auto-discovered by pytest due to naming
 
 ## 0.2.0 (2023-06-07)